Reorganize ack,businessData,reminder,style ui services
