+++ ui/main/src/app/business/application-loader.ts
@@ -14,16 +14,16 @@ import {RemoteLoggerService} from '../services/logs/RemoteLoggerService';
 import {I18nService} from './services/translation/i18n.service';
 import {UserService} from './services/users/user.service';
 import {OpfabAPI} from '../api/opfab.api';
-import {GlobalStyleService} from './services/global-style.service';
+import {GlobalStyleService} from '../services/style/global-style.service';
 import {EntitiesService} from './services/users/entities.service';
 import {GroupsService} from './services/users/groups.service';
 import {PerimetersService} from './services/users/perimeters.service';
 import {ProcessesService} from './services/businessconfig/processes.service';
 import {OpfabEventStreamService} from './services/events/opfabEventStream.service';
-import {AcknowledgeService} from './services/acknowledge.service';
+import {AcknowledgeService} from '../services/acknowlegment/AcknowledgeService';
 import {DateTimeFormatterService} from '../services/dateTimeFormatter/DateTimeFormatterService';
 import {AdminProcessesService} from './services/businessconfig/adminprocess.service';
-import {BusinessDataService} from './services/businessconfig/businessdata.service';
+import {BusinessDataService} from '../services/businessdata/businessdata.service';
 import {CardService} from './services/card/card.service';
 import {SupervisedEntitiesService} from './services/users/supervised-entities.service';
 import {SelectedCardLoaderService} from './services/card/selectedCardLoader.service';
