+++ ui/main/src/app/business/services/events/application-update.service.ts
@@ -18,7 +18,7 @@ import {Utilities} from '../../common/utilities';
 import {ApplicationEventsService} from './application-events.service';
 import {OpfabEventStreamService} from './opfabEventStream.service';
 import {ProcessesService} from '../businessconfig/processes.service';
-import {BusinessDataService} from '../businessconfig/businessdata.service';
+import {BusinessDataService} from '../../../services/businessdata/businessdata.service';
 import {ConfigService} from '../../../services/config/ConfigService';
 
 export class ApplicationUpdateService {
