+++ ui/main/src/app/business/view/navbar/navbarMenu_OnMenuClick.view.spec.ts
@@ -17,7 +17,7 @@ import {NavbarMenuView} from './navbarMenu.view';
 import {MenuEntryLinkTypeEnum} from '@ofModel/menu.model';
 import {ApplicationRouterMock} from '@tests/mocks/applicationRouter.mock';
 import {RouterService} from 'app/business/services/router.service';
-import {GlobalStyleService} from 'app/business/services/global-style.service';
+import {GlobalStyleService} from '@ofServices/style/global-style.service';
 
 declare const opfabStyle;
