+++ ui/main/src/app/modules/core/application-loading/application-loading.component.ts
@@ -16,8 +16,8 @@ import {OpfabEventStreamServer} from 'app/business/server/opfabEventStream.serve
 import {CurrentUserStore} from 'app/business/store/current-user.store';
 import {AuthService} from 'app/authentication/auth.service';
 import {AuthenticationMode} from 'app/authentication/auth.model';
-import {RemoteLoggerServer} from 'app/services/logs/RemoteLoggerServer';
-import {ConfigServer} from 'app/services/config/ConfigServer';
+import {RemoteLoggerServer} from '@ofServices/logs/server/RemoteLoggerServer';
+import {ConfigServer} from '@ofServices/config/server/ConfigServer';
 import {ApplicationLoader} from 'app/business/application-loader';
 import {TranslationService} from 'app/business/services/translation/translation.service';
 import {UserServer} from 'app/business/server/user.server';
@@ -26,9 +26,9 @@ import {EntitiesServer} from 'app/business/server/entities.server';
 import {GroupsServer} from 'app/business/server/groups.server';
 import {PerimetersServer} from 'app/business/server/perimeters.server';
 import {ProcessServer} from 'app/business/server/process.server';
-import {AcknowledgeServer} from 'app/business/server/acknowledge.server';
+import {AcknowledgeServer} from '@ofServices/acknowlegment/server/AcknowledgeServer';
 import {AdminProcessServer} from 'app/business/server/adminprocess.server';
-import {BusinessDataServer} from 'app/business/server/businessData.server';
+import {BusinessDataServer} from '@ofServices/businessdata/server/BusinessDataServer';
 import {CardServer} from 'app/business/server/card.server';
 import {SupervisedEntitiesServer} from 'app/business/server/supervised-entities.server';
 import {ExternalDevicesServer} from '@ofServices/notifications/server/ExternalDevicesServer';
