+++ ui/main/src/app/services/config/server/ConfigServer.ts
@@ -7,11 +7,11 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {MonitoringConfig} from './model/MonitoringConfig';
-import {RealTimeScreens} from './model/RealTimeScreensConfig';
+import {MonitoringConfig} from '../model/MonitoringConfig';
+import {RealTimeScreens} from '../model/RealTimeScreensConfig';
 import {Observable} from 'rxjs';
-import {ServerResponse} from '../../business/server/serverResponse';
-import {ProcessMonitoringConfig} from './model/ProcessMonitoringConfig';
+import {ServerResponse} from '../../../business/server/serverResponse';
+import {ProcessMonitoringConfig} from '../model/ProcessMonitoringConfig';
 
 export abstract class ConfigServer {
     abstract getWebUiConfiguration(): Observable<ServerResponse<any>>;
