+++ ui/main/src/app/services/logs/server/RemoteLoggerServer.ts
