+++ ui/main/src/app/services/reminder/ReminderUtils.ts
