+++ ui/main/src/main.ts
@@ -15,8 +15,8 @@ import {LoggerService as logger} from 'app/services/logs/LoggerService';
 import {LocationStrategy, HashLocationStrategy, CommonModule} from '@angular/common';
 import {HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
 import {TokenInjector} from 'app/server/interceptors.service';
-import {AcknowledgeServer} from './app/business/server/acknowledge.server';
-import {AngularAcknowledgeServer} from './app/server/angularAcknowledgement.server';
+import {AcknowledgeServer} from './app/services/acknowlegment/server/AcknowledgeServer';
+import {AngularAcknowledgeServer} from './app/services/acknowlegment/server/AngularAcknowledgementServer';
 import {EntitiesServer} from './app/business/server/entities.server';
 import {AngularEntitiesServer} from './app/server/angularEntities.server';
 import {SupervisedEntitiesServer} from './app/business/server/supervised-entities.server';
@@ -31,16 +31,16 @@ import {UserActionLogsServer} from './app/business/server/user-action-logs.serve
 import {AngularUserActionLogsServer} from './app/server/angularUser-Action-Logs.server';
 import {AdminProcessServer} from './app/business/server/adminprocess.server';
 import {AngularAdminProcessesServer} from './app/server/angularAdminProcess.server';
-import {RemoteLoggerServer} from './app/services/logs/RemoteLoggerServer';
-import {AngularRemoteLoggerServer} from './app/services/logs/AngularRemoteLoggerServer';
-import {ConfigServer} from './app/services/config/ConfigServer';
-import {AngularConfigServer} from './app/services/config/AngularConfigServer';
+import {RemoteLoggerServer} from './app/services/logs/server/RemoteLoggerServer';
+import {AngularRemoteLoggerServer} from './app/services/logs/server/AngularRemoteLoggerServer';
+import {ConfigServer} from './app/services/config/server/ConfigServer';
+import {AngularConfigServer} from './app/services/config/server/AngularConfigServer';
 import {TemplateCssServer} from './app/business/server/template-css.server';
 import {AngularTemplateCssServer} from './app/server/angularTemplate-css.service';
 import {ProcessServer} from './app/business/server/process.server';
 import {AngularProcessServer} from './app/server/angularProcess.server';
-import {BusinessDataServer} from './app/business/server/businessData.server';
-import {AngularBusinessDataServer} from './app/server/angularBusinessData.server';
+import {BusinessDataServer} from './app/services/businessdata/server/BusinessDataServer';
+import {AngularBusinessDataServer} from './app/services/businessdata/server/AngularBusinessData.server';
 import {SettingsServer} from './app/business/server/settings.server';
 import {AngularSettingsServer} from './app/server/angularSettings.server';
 import {OpfabEventStreamServer} from './app/business/server/opfabEventStream.server';
